break out ProjectSerializer
